refactor: move types to types.ts
