--- src/stories/components/Forms/Select/types/Option.ts
@@ -1,6 +1,0 @@
-import { MenuItemProps } from '@mui/material/MenuItem';
-
-export type Option = {
-  value: MenuItemProps['value'];
-  label: string;
-};
